IGNITE-19566 Rename TraceIdGenerator to TraceIdUtils
